// web/src/lib/api/generated/models/tokenOut.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { TokenOutState } from './tokenOutState.ts'
import type { TokenScopeOut } from './tokenScopeOut.ts'

export interface TokenOut {
  actor: string
  created_at: string
  expires_at?: string
  id: number
  /** @nullable */
  last_used_at: string | null
  /** @nullable */
  name: string | null
  /** @nullable */
  revoked_at: string | null
  scope?: TokenScopeOut
  state: TokenOutState
  [key: string]: unknown
}